requirements.txt
# FreightMetrics MVP - Dependencias
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
pandas>=1.5.0
plotly>=5.0.0
requests>=2.31.0
PyYAML>=6.0
streamlit-authenticator>=0.2.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Google Services
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0

# Opcional para mejor rendimiento
numpy>=1.24.0